Fixes #3294: space for the periods has already been allocated by ompi_comm_split(), and the entire set of periods from the old communicator have already been copied to the new communicator.  But up here in mca_topo_base_cart_sub(), we need to subset the periods that are actually stored on the new communicator according to remain_dims (just like we did for the set of dimensions).
This commit renames a few variables to be a little less misleading,
and then adds a loop to copy over the periods information.  I could
have added this into the first loop (that subset-copies the
dimensions), but this code is already confusing enough and this is not
a performance-critical section: so I made it a new loop.

Note that all the topo code will be revamped a bit when the new
MPI-2.2 topo stuff (currently off in a mercurial branch) finally makes
it back to the SVN trunk.  But that new stuff will only get to v1.7 --
this commit will need to be CMR'ed to v1.6.x.

cmr:v1.7
cmr:v1.6.2