Fixes #3306: CMR String NULL reference protection in C/R code path (SSTORE)
---svn-pre-commit-ignore-below---

r27289 [[BR]]
Protect a potentially uninitialized variable (orte_sstore_base_global_snapshot_ref).

If orte_sstore_base_global_snapshot_ref is null, then it will default appropriately when it is used. When prelaunching we always specify this parameter, but if we are not prelaunching it is possible to allow this to be null and it will initialize when used. However we setup the prelaunching variable in both situtations and in the latter that would result in a NULL reference. This patch protects that code segment.